refactor(api): split test assertions from server harness
